requirements.txt



langchain==0.3.25
langchain-anthropic==0.3.15
langchain-community==0.3.25
langgraph==0.2.74              # Python 3.13 support added in 0.2.28

# ── Vector database ───────────────────────────────────────────────────────────
# IMPORTANT: pinecone-client is deprecated. Use pinecone instead.
# Run: pip uninstall pinecone-client before installing this
pinecone==6.0.1                # renamed from pinecone-client, Python 3.13 tested

# ── Retrieval ─────────────────────────────────────────────────────────────────
rank-bm25==0.2.2               # no Python 3.13 issues, stable

# ── Reranking ─────────────────────────────────────────────────────────────────
cohere==5.15.0

# ── Spotify API ───────────────────────────────────────────────────────────────
spotipy==2.25.0

# ── Audio analysis ────────────────────────────────────────────────────────────
librosa==0.10.2.post1          # works with numpy 2.x, Python 3.13 compatible
numpy>=2.0.0                   # CRITICAL: numpy 1.x has NO Python 3.13 support
soundfile==0.12.1

# ── MERT audio embeddings ─────────────────────────────────────────────────────
torch>=2.0.0
transformers>=4.35.0

# ── Evaluation ────────────────────────────────────────────────────────────────
ragas==0.2.15
datasets==3.3.2

# ── Observability ─────────────────────────────────────────────────────────────
# NOTE: langfuse 3.x changed import path
# Old: from langfuse.callback import CallbackHandler
# New: from langfuse.langchain import CallbackHandler
langfuse==3.0.3

# ── API server ────────────────────────────────────────────────────────────────
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.1.0
pydantic==2.11.3
httpx==0.28.1
jinja2==3.1.6

# ── ML models ─────────────────────────────────────────────────────────────
scikit-learn>=1.3.0
joblib>=1.3.0
xgboost>=2.0.0